@@ -7,11 +7,17 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@types/jest": "^28.1.1",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.13",
+    "@types/react-dom": "^18.0.5",
+    "axios": "^0.27.2",
     "gh-pages": "^2.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "3.4.1",
-    "tachyons": "^4.12.0"
+    "tachyons": "^4.12.0",
+    "typescript": "^3.9.10"
   },
   "scripts": {
     "start": "react-scripts start",
